bot/requirements.txt
python-telegram-bot
flask
gunicorn
requests
yookassa
